refactor: extract tokenCrypto from useGitHubSync
Move encryptToken/decryptToken to dedicated module with focused
unit tests covering roundtrip, wrong passphrase, IV uniqueness,
and tampered IV rejection.
