+++ src/hooks/useGitHubSync.ts
@@ -1,7 +1,7 @@
 import { useState, useEffect, useRef, useCallback } from 'react'
-import { deriveKey, bytesToB64, b64ToBytes } from '../utils/crypto'
 import { getStorageItem, setStorageItem } from '../utils/storage'
 import { toBase64, fromBase64 } from './base64Utils'
+import { encryptToken, decryptToken } from './tokenCrypto'
 
 export { toBase64, fromBase64 } from './base64Utils'
 
@@ -70,36 +70,6 @@ export const loadConfig = (): GitHubSyncConfig => {
   } catch {
     return DEFAULT_CONFIG
   }
-}
-
-const encryptToken = async (
-  token: string,
-  passphrase: string,
-): Promise<{ encryptedToken: string; tokenSalt: string; tokenIv: string }> => {
-  const salt = crypto.getRandomValues(new Uint8Array(16))
-  const iv = crypto.getRandomValues(new Uint8Array(12))
-  const key = await deriveKey(passphrase, salt)
-  const ciphertext = await crypto.subtle.encrypt({ name: 'AES-GCM', iv }, key, new TextEncoder().encode(token))
-  return {
-    encryptedToken: bytesToB64(new Uint8Array(ciphertext)),
-    tokenSalt: bytesToB64(salt),
-    tokenIv: bytesToB64(iv),
-  }
-}
-
-const decryptToken = async (
-  encryptedToken: string,
-  passphrase: string,
-  tokenSalt: string,
-  tokenIv: string,
-): Promise<string> => {
-  const key = await deriveKey(passphrase, b64ToBytes(tokenSalt))
-  const plaintext = await crypto.subtle.decrypt(
-    { name: 'AES-GCM', iv: b64ToBytes(tokenIv) },
-    key,
-    b64ToBytes(encryptedToken),
-  )
-  return new TextDecoder().decode(plaintext)
 }
 
 export const useGitHubSync = () => {
